requirements.txt
psycopg2
python-decouple~=3.4

